pyproject.toml
[project]
name = "a-frame-memory"
version = "0.1.0"
description = "A-Frame: Training-free Agentic Cognitive Memory Framework"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
dependencies = [
    "langgraph>=0.4.0",
    "langchain-core>=0.3.0",
    "fastapi[all]>=0.115.0",
    "openai>=1.77.0",
    "google-genai>=1.0.0",
    "pydantic-settings>=2.0.0",
    "networkx>=3.2",
    "tiktoken>=0.9.0",
    "pyyaml>=6.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
production = [
    "neo4j>=5.20.0",
    "lancedb>=0.20.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["a_frame"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
